has leaks on tests 1 && 3; pass tests 1 && 3; test 2 fail

+++ Project2/ServerPart/src/server/mylib.c
@@ -1,6 +1,6 @@
 #include "mylib.h"
 
-void process_message(struct message * sms) {
+void process_message(struct message* sms) {
     usleep(sms->tskload * 10e3);
-    sms->tskres = sms->rid * 100;
+    sms->tskres = (int)sms->rid * 100;
 }
